Integrate PageSectionTabs component into GuestProfile for improved tab navigation and section descriptions
